Move Context-related methods to 1.8+-specific files (with build tags)
