+++ squirrel.go
@@ -5,7 +5,6 @@ package squirrel
 
 import (
 	"bytes"
-	"context"
 	"database/sql"
 	"fmt"
 	"strings"
@@ -19,30 +18,6 @@ import (
 // as passed to e.g. database/sql.Exec. It can also return an error.
 type Sqlizer interface {
 	ToSql() (string, []interface{}, error)
-}
-
-// Execer is the interface that wraps the Exec method.
-//
-// Exec executes the given query as implemented by database/sql.Exec.
-type Execer interface {
-	Exec(query string, args ...interface{}) (sql.Result, error)
-	ExecContext(ctx context.Context, query string, args ...interface{}) (sql.Result, error)
-}
-
-// Queryer is the interface that wraps the Query method.
-//
-// Query executes the given query as implemented by database/sql.Query.
-type Queryer interface {
-	Query(query string, args ...interface{}) (*sql.Rows, error)
-	QueryContext(ctx context.Context, query string, args ...interface{}) (*sql.Rows, error)
-}
-
-// QueryRower is the interface that wraps the QueryRow method.
-//
-// QueryRow executes the given query as implemented by database/sql.QueryRow.
-type QueryRower interface {
-	QueryRow(query string, args ...interface{}) RowScanner
-	QueryRowContext(ctx context.Context, query string, args ...interface{}) RowScanner
 }
 
 // BaseRunner groups the Execer and Queryer interfaces.
@@ -67,20 +42,12 @@ func (r *dbRunner) QueryRow(query string, args ...interface{}) RowScanner {
 	return r.DB.QueryRow(query, args...)
 }
 
-func (r *dbRunner) QueryRowContext(ctx context.Context, query string, args ...interface{}) RowScanner {
-	return r.DB.QueryRowContext(ctx, query, args...)
-}
-
 type txRunner struct {
 	*sql.Tx
 }
 
 func (r *txRunner) QueryRow(query string, args ...interface{}) RowScanner {
 	return r.Tx.QueryRow(query, args...)
-}
-
-func (r *txRunner) QueryRowContext(ctx context.Context, query string, args ...interface{}) RowScanner {
-	return r.Tx.QueryRowContext(ctx, query, args...)
 }
 
 func setRunWith(b interface{}, baseRunner BaseRunner) interface{} {
@@ -124,30 +91,6 @@ func QueryWith(db Queryer, s Sqlizer) (rows *sql.Rows, err error) {
 func QueryRowWith(db QueryRower, s Sqlizer) RowScanner {
 	query, args, err := s.ToSql()
 	return &Row{RowScanner: db.QueryRow(query, args...), err: err}
-}
-
-// ExecContextWith ExecContexts the SQL returned by s with db.
-func ExecContextWith(ctx context.Context, db Execer, s Sqlizer) (res sql.Result, err error) {
-	query, args, err := s.ToSql()
-	if err != nil {
-		return
-	}
-	return db.ExecContext(ctx, query, args...)
-}
-
-// QueryContextWith QueryContexts the SQL returned by s with db.
-func QueryContextWith(ctx context.Context, db Queryer, s Sqlizer) (rows *sql.Rows, err error) {
-	query, args, err := s.ToSql()
-	if err != nil {
-		return
-	}
-	return db.QueryContext(ctx, query, args...)
-}
-
-// QueryRowContextWith QueryRowContexts the SQL returned by s with db.
-func QueryRowContextWith(ctx context.Context, db QueryRower, s Sqlizer) RowScanner {
-	query, args, err := s.ToSql()
-	return &Row{RowScanner: db.QueryRowContext(ctx, query, args...), err: err}
 }
 
 // DebugSqlizer calls ToSql on s and shows the approximate SQL to be executed
