+++ squirrel_test.go
@@ -6,8 +6,6 @@ import (
 	"strings"
 	"testing"
 
-	"context"
-
 	"github.com/stretchr/testify/assert"
 )
 
@@ -48,24 +46,6 @@ func (s *DBStub) Query(query string, args ...interface{}) (*sql.Rows, error) {
 }
 
 func (s *DBStub) QueryRow(query string, args ...interface{}) RowScanner {
-	s.LastQueryRowSql = query
-	s.LastQueryRowArgs = args
-	return &Row{RowScanner: &RowStub{}}
-}
-
-func (s *DBStub) ExecContext(ctx context.Context, query string, args ...interface{}) (sql.Result, error) {
-	s.LastExecSql = query
-	s.LastExecArgs = args
-	return nil, nil
-}
-
-func (s *DBStub) QueryContext(ctx context.Context, query string, args ...interface{}) (*sql.Rows, error) {
-	s.LastQuerySql = query
-	s.LastQueryArgs = args
-	return nil, nil
-}
-
-func (s *DBStub) QueryRowContext(ctx context.Context, query string, args ...interface{}) RowScanner {
 	s.LastQueryRowSql = query
 	s.LastQueryRowArgs = args
 	return &Row{RowScanner: &RowStub{}}
