+++ stmtcacher.go
@@ -1,7 +1,6 @@
 package squirrel
 
 import (
-	"context"
 	"database/sql"
 	"sync"
 )
@@ -70,30 +69,6 @@ func (sc *stmtCacher) QueryRow(query string, args ...interface{}) RowScanner {
 		return &Row{err: err}
 	}
 	return stmt.QueryRow(args...)
-}
-
-func (sc *stmtCacher) ExecContext(ctx context.Context, query string, args ...interface{}) (res sql.Result, err error) {
-	stmt, err := sc.Prepare(query)
-	if err != nil {
-		return
-	}
-	return stmt.ExecContext(ctx, args...)
-}
-
-func (sc *stmtCacher) QueryContext(ctx context.Context, query string, args ...interface{}) (rows *sql.Rows, err error) {
-	stmt, err := sc.Prepare(query)
-	if err != nil {
-		return
-	}
-	return stmt.QueryContext(ctx, args...)
-}
-
-func (sc *stmtCacher) QueryRowContext(ctx context.Context, query string, args ...interface{}) RowScanner {
-	stmt, err := sc.Prepare(query)
-	if err != nil {
-		return &Row{err: err}
-	}
-	return stmt.QueryRowContext(ctx, args...)
 }
 
 type DBProxyBeginner interface {
